+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## dev-master
 
-* Added `check-environment.php` script to report on the environment in which script is running.
+* Added image merging demo `/demo/merge-document-image.php`.
+* Added `/bin/check-environment.php` script to report on the environment in which script is running.
 * Added documentation.
 * Added a basic `mergeDocument()` demo.